requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
httpx==0.27.0
apscheduler==3.10.4
loguru==0.7.2
beautifulsoup4==4.12.3
lxml==6.0.2
requests==2.32.3
playwright==1.48.0
feedparser==6.0.11
sqlalchemy==2.0.35
aiosqlite==0.20.0

